requirements.txt
python-binance
ta-lib
numpy
websocket_client